using vps to download contents

+++ ftp.py
@@ -1,5 +1,5 @@
 from ftplib import FTP
-import config, constants
+import config, constants, os
 
 def download(filename):
     config_data = config.get_config()
@@ -7,8 +7,10 @@ def download(filename):
     ftp = FTP(config_data['server_ip'])
     ftp.login(config_data['username'], config_data['password'])
     ftp.cwd(constants.FTP_DIR)
-    with open(filename, 'wb') as f:
-        ftp.retrbinary('RETR %s' % filename, f.write)
-        print('finished downloading %s' % filename)
+
+    if not os.path.exists(filename):
+        with open(filename, 'wb') as f:
+            ftp.retrbinary('RETR %s' % filename, f.write)
+            print('finished downloading %s' % filename)
 
     ftp.quit()
